requirements.txt
absl-py==2.3.1
annotated-types==0.7.0
anyio==4.10.0
attrs==25.3.0
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
contourpy==1.3.2
cycler==0.12.1
dotenv==0.9.9
ecdsa==0.19.1
exceptiongroup==1.3.0
fastapi==0.116.1
flatbuffers==25.2.10
fonttools==4.59.1
greenlet==3.2.4
idna==3.10
jax==0.6.2
jaxlib==0.6.2
kiwisolver==1.4.9
matplotlib==3.10.5
mediapipe==0.10.21
ml_dtypes==0.5.3
numpy==1.26.4
opencv-contrib-python==4.11.0.86
opencv-python==4.11.0.86
opt_einsum==3.4.0
packaging==25.0
passlib==1.7.4
pillow==11.3.0
protobuf==4.25.8
psutil==7.0.0
psycopg2-binary==2.9.10
pyasn1==0.6.1
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pyparsing==3.2.3
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.17.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-jose==3.5.0
requests==2.32.4
rsa==4.9.1
scipy==1.15.3
sentencepiece==0.2.1
six==1.17.0
sniffio==1.3.1
sounddevice==0.5.2
SQLAlchemy==2.0.43
starlette==0.47.2
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
